doc: mypage routes 활용 기능 구현중
